// src/store/user/user.types.ts
export enum UserActionsTypes {
  setCurrentUser = "user/SET_CURRENT_USER",
  checkUserSession = "user/CHECK_USER_SESSION",
  googleSignInStart = "user/GOOGLE_SIGN_IN_START",
  emailSignInStart = "user/EMAIL_SIGN_IN_START",
  signInSuccess = "user/SIGN_IN_SUCCESS",
  signInFailure = "user/SIGN_IN_FAILURE",
  signOutFailure = "user/SIGN_OUT_FAILURE",
  signOutSuccess = "user/SIGN_OUT_SUCCESS",
  signOutStart = "user/SIGN_OUT_START",
  signUpFailure = "user/SIGN_UP_FAILURE",
  signUpSuccess = "user/SIGN_UP_SUCCESS",
  signUpStart = "user/SIGN_UP_START",
}